requirements.txt
# ── Core pipeline ────────────────────────────────────────────────────────
faster-whisper>=1.0.0
TTS>=0.22.0
pydub>=0.25.1
pyannote.audio>=4.0.0

# ── Translation (NLLB offline fallback) ─────────────────────────────────
transformers>=4.38.0
sentencepiece>=0.2.0
protobuf>=4.25.0

# ── GUI ──────────────────────────────────────────────────────────────────
gradio>=4.0.0

# ── Utilities ────────────────────────────────────────────────────────────
requests>=2.31.0
soundfile>=0.12.1
numpy>=1.24.0
